requirements.txt
asgiref==3.4.1
Django==3.2.5
django-extensions==3.1.3
djangorestframework==3.12.4
ngrok==0.0.1
pygraphviz==1.7
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
